Fixed incorrect indentation with hash marks in resources.
The problem was that we marked the hash mark as comment character,
which means that syntax-ppss considers everything until the comment
end character as a comment.

However, hash marks appear regularly in URLs. This means that
syntax-ppss gets confused. Using modify-syntax to set the sequence
" #" to a comment character instead, but didn't.

I've replaced comment and string fontification with regexes, and
disabled all syntax-ppss comment parsing.

This will now break if metacharacters are used in a comment. Let's see
if we can excise remaining instances of syntax-ppss.