Restructured indentation to be less complicated.
The functionalities for all the "block" types were the same, so they've been merged with each other.
Unfortunately, this has exposed a small bug with multiple open parentheses on a single line. The
first line of a graph (for example) will indent correctly, since it takes the previous line's
indentation, while the next line is based on the parenthesis count and will be indented one indent
width further.

I could fix this by always indenting off the previous line, instead of using the parenthesis depth.